fullstack_application/requirements.txt
argcomplete==1.12.2
asgiref==3.2.7
aws-psycopg2==1.2.1
boto3==1.17.42
botocore==1.20.42
certifi==2020.12.5
cffi==1.14.0
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==3.10.0
jmespath==0.10.0
kappa==0.6.0
pep517==0.10.0
pip-tools==6.0.1
placebo==0.9.0
pycparser==2.20
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.59.0
troposphere==2.7.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.1
